move expoted common interface to separat file
